+++ controllers/ExpenseController.php
@@ -190,8 +190,43 @@ class ExpenseController extends Controller
                 return $this->redirect(['view', 'id' => $model->id]);
         }
 
+        // Get a list of all user-assigned cost projects
+        $costprojects = Costproject::find()
+        ->select(['costproject.*'])
+        ->innerJoinWith('users')
+        ->where(['user.id' => Yii::$app->user->id])
+        ->all();
+   
+        // Get a list of all participants, if a cost project is already selected
+        $participants = null;
+        if(!empty($model->costprojectId))
+            $participants = Costproject::findOne(['id' => $model->costprojectId])->getParticipantsList();
+        if(!is_array($model->participants)) {
+            if(empty($model->participants))
+                $model->participants = array();
+            else
+                $model->participants = explode(';', $model->participants);
+        }
+
+        // Get all titles
+        $costprojectIDs = Costproject::find()
+            ->select(['costproject.id'])
+            ->innerJoinWith('users')
+            ->where(['user.id' => Yii::$app->user->id])
+            ->column();
+
+        $titles = Expense::find()
+            ->select('title')
+            ->where(['costprojectId' => $costprojectIDs])
+            ->groupBy('title')
+            ->orderBy('title ASC')
+            ->column();
+
         return $this->render('update', [
-            'model' => $model,
+            'model'         => $model,
+            'costprojects'  => $costprojects,
+            'participants'  => $participants,
+            'titles'        => $titles,
         ]);
     }