+++ views/costproject/_view.php
@@ -30,6 +30,28 @@ use app\dictionaries\CurrencyCodesDict;
             'attribute' => 'useCurrency',
             'format' => 'checkbox',
         ],
+        [
+            'attribute' => Yii::t('app', 'Users'),
+            'format' => 'raw',
+            'value' => function($data) {
+                $tmp = [];
+                foreach($data->users as $user) {
+                    $item =  $user->displayName.' (#'.$user->id.')';
+                    if((int)$user->id!==(int)Yii::$app->user->id && Yii::$app->user->can('updateCostproject', ['costproject'=>$data]))
+                        $item .= ' ' . Html::a(Html::icon('trash-2'), ['remove-user', 'AddUserForm[costprojectId]'=>$data->id, 'AddUserForm[username]' => $user->username], [
+                            'class' => 'btn btn-primary btn-sm',
+                            'data' => [
+                                'confirm' => Yii::t('app', 'Are you sure you want to remove this user?'),
+                                'method' => 'post',
+                            ]
+                        ]);
+                    $tmp[] = $item;
+                }
+                if($this->context->action->id!=='manage-users' && Yii::$app->user->can('updateCostproject', ['costproject'=>$data]))
+                    $tmp[] = Html::a(Html::icon('plus-square') . Yii::t('app', 'Manage Users'), ['manage-users', 'id'=>$data->id], ['class' => 'btn btn-sm btn-primary mt-2']);
+                return join('<br>', $tmp);
+            }
+        ],
         // 'id',
     ],
 ]) ?>