+++ views/expense/create.php
@@ -21,10 +21,10 @@ $this->params['breadcrumbs'][] = $this->title;
     <h1><?= Html::encode($this->title) ?></h1>
 
     <?= $this->render('_form', [
-        'model' => $model,
-        'costprojects' => $costprojects,
-        'participants' => $participants,
-        'titles' => $titles,
+        'model'         => $model,
+        'costprojects'  => $costprojects,
+        'participants'  => $participants,
+        'titles'        => $titles,
     ]) ?>
 
 </div>